+++ GPTifier/src/interface/command_fine_tune.cpp
@@ -124,6 +124,43 @@ void delete_ft_model(int argc, char **argv)
 
 // Print fine tuning jobs -----------------------------------------------------------------------------------
 
+ParamsGetFineTuningJobs read_cli_list_ft_jobs(int argc, char **argv)
+{
+    ParamsGetFineTuningJobs params;
+
+    while (true) {
+        static struct option long_options[] = {
+            { "help", no_argument, 0, 'h' },
+            { "json", no_argument, 0, 'j' },
+            { "limit", required_argument, 0, 'l' },
+            { 0, 0, 0, 0 },
+        };
+
+        int option_index = 0;
+        int c = getopt_long(argc, argv, "hjl:", long_options, &option_index);
+
+        if (c == -1) {
+            break;
+        }
+
+        switch (c) {
+            case 'h':
+                cli::help_command_fine_tune_list_jobs();
+                exit(EXIT_SUCCESS);
+            case 'j':
+                params.print_raw_json = true;
+                break;
+            case 'l':
+                params.limit = optarg;
+                break;
+            default:
+                cli::exit_on_failure();
+        }
+    };
+
+    return params;
+}
+
 void print_ft_jobs(const FineTuningJobs &ft_jobs)
 {
     print_sep();
@@ -140,7 +177,7 @@ void print_ft_jobs(const FineTuningJobs &ft_jobs)
 
 void list_ft_jobs(int argc, char **argv)
 {
-    ParamsGetFineTuningJobs params = cli::get_opts_get_fine_tuning_jobs(argc, argv);
+    ParamsGetFineTuningJobs params = read_cli_list_ft_jobs(argc, argv);
     std::string limit = params.limit.value_or("20");
 
     FineTuningJobs ft_jobs = get_fine_tuning_jobs(limit);
